Template part 2: In prover.rs, initialization of the state with the initial value and the binary decomposition, with one bit per column. Then, the update consists in either computing the binary decomposition (with the first column still containing the current value) or the Collatz update. During the Collatz update, the rest of the columns are filled with a value that will reduce the likelihood of all-zero columns.
In air.rs: we add 129 periodic degrees (3, with cycle size=2). There are 129 constraints. When checking the binary decomposition, we compute the value based on the binary decomposition, and check whether the result equals the stored value. We also check that the next state's and the current state's first values concur. In the Collatz checking step, we ensure that the current state's binary decomposition only contains binary values, and check whether the next state's first value is equal to a Collatz update from the current state's first value.

We also consider two assertions, concerning the first and last value of the Collatz sequence.

With these changes, the proof is generated but for some sequence length values, I get the error: "Failed to verify proof: constraint evaluations over the out-of-domain frame are inconsistent"